+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+- Added A1Range and A1Reference examples to readme and example
+- Adjusted area to be a double for A1 and build this into a range
+- Fixed minor bug with parsing a partial range in a reference
+- Added github action build pipeline with coverage
+- Move test coverage to 100%
+
 ## 1.0.1
 
 - Minor fixes in docs and imports